+++ R/mod_afs_study_beach_census.R
@@ -425,7 +425,7 @@ mod_afs_study_beach_census_server <- function(id, pool, season.df) {
 
       #-------------------------------------------------------------------------
       ### Send off
-      observe(mod_output_server("out", session, tbl_output, plot_output))
+      observe(mod_output_server("out", tbl_output, plot_output))
     }
   )
 }