+++ R/mod_ccamlr_pup_weights.R
@@ -243,7 +243,7 @@ mod_ccamlr_pup_weights_server <- function(id, pool, season.df) {
 
       #-------------------------------------------------------------------------
       ### Send off
-      observe(mod_output_server("out", session, tbl_output, plot_output))
+      observe(mod_output_server("out", tbl_output, plot_output))
     }
   )
 }